strategy pattern in reconciling
On-behalf-of: @SAP a.shcherbatiuk@sap.com
